+++ MercadoPagoSDK/MercadoPagoSDK/PXBusinessResultViewModel.swift
@@ -7,8 +7,16 @@
 //
 
 import UIKit
+import MercadoPagoPXTracking
 
 class PXBusinessResultViewModel: NSObject, PXResultViewModelInterface {
+    var screenName: String { return TrackingUtil.SCREEN_NAME_PAYMENT_RESULT }
+    var screenId: String { return TrackingUtil.SCREEN_ID_PAYMENT_RESULT_BUSINESS }
+    
+    func trackInfo() {
+        MPXTracker.sharedInstance.trackScreen(screenId: screenId, screenName: screenName, properties: [String:String]())
+    }
+    
 
     let businessResult: PXBusinessResult
     let paymentData: PaymentData
